+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Please add your entries according to this format.
 
 ## Unreleased
 
+- `--include-only` now works with paths that are file-separator independent.
 - Added support for Gradle 7.0+
 - Added task ordering between `ktfmt*` tasks and `compileKotlin` tasks. This fix the correctness warning introduced with Gradle 7.0